test: move fixture to conftest.py
