+++ tests/domain/test_model.py
@@ -7,51 +7,6 @@ from stone.domain import commands, events, model
 from stone.domain.model import BattleField, Minion, Player
 
 
-@dataclass
-class TestMin(model.Minion):
-    name: str = "test_minion"
-    attack: int = 1
-    life: int = 100
-    data: str = field(default="sss")
-
-
-@pytest.fixture
-def minions():
-    minions = [
-        # Minion(attack=1, life=100, name="test_min", mana=1) for _ in range(2)
-        TestMin("dd"),
-        TestMin("bb"),
-    ]
-    return minions
-
-
-@pytest.fixture
-def players(minions):
-    player1 = Player()
-    player2 = Player()
-    player1.minion_field.append(minions[0])
-    player2.minion_field.append(minions[1])
-    return player1, player2
-
-
-@pytest.fixture
-def battle_field(players) -> BattleField:
-    player_dict = {player.uuid: player for player in players}
-    field = BattleField(players=player_dict)
-    return field
-
-
-@pytest.fixture
-def message_bus(battle_field):
-    message_bus = bootstrap(battle_field)
-    return message_bus
-
-
-@pytest.fixture
-def spell_with_attack_3():
-    return model.Spell(attack=3, name="test_spell")
-
-
 def test_minion_melee_attack_player_with_damage_1_decrease_1_life(
     battle_field, message_bus
 ):
